Inventory from main

+++ install/install_monitoring.sh
@@ -8,6 +8,28 @@ echo "########################################################"
 
 install_monitoring () {
 
+  echo "### Which cluster you wnat to monitor? ###"
+  select cluster in "mainnet-beta" "testnet"; do
+      case $cluster in
+          mainnet-beta ) inventory="mainnet.yaml"; break;;
+          testnet ) inventory="testnet.yaml"; break;;
+      esac
+  done
+
+
+  echo "### Please type your validator name: "
+  read VALIDATOR_NAME
+  echo "### Please type the full path to your validator keys: "
+  read PATH_TO_VALIDATOR_KEYS
+
+  if [ ! -f "$PATH_TO_VALIDATOR_KEYS/validator-keypair.json" ]
+  then
+    echo "key $PATH_TO_VALIDATOR_KEYS/validator-keypair.json not found. Pleas verify and run the script again"
+    exit
+  fi
+
+  read -e -p "### Please tell which user is running validator: " SOLANA_USER
+  cd
   rm -rf sv_manager/
 
   if [[ $(which apt | wc -l) -gt 0 ]]
@@ -38,45 +60,22 @@ install_monitoring () {
   cd ./sv_manager || exit
   cp -r ./inventory_example ./inventory
 
-  echo "### Which cluster do you want to monitor? ###"
-  select cluster in "mainnet-beta" "testnet"; do
-      case $cluster in
-          mainnet-beta ) cluster_environment="mainnet-beta"; break;;
-          testnet ) cluster_environment="testnet"; break;;
-      esac
-  done
-
-
-  echo "### Please type your validator name: "
-  read VALIDATOR_NAME
-  echo "### Please type the full path to your validator keys: "
-  read PATH_TO_VALIDATOR_KEYS
-
-  if [ ! -f "$PATH_TO_VALIDATOR_KEYS/validator-keypair.json" ]
-  then
-    echo "key $PATH_TO_VALIDATOR_KEYS/validator-keypair.json not found. Pleas verify and run the script again"
-    exit
-  fi
-
-  read -e -p "### Please tell which user is running validator: " SOLANA_USER
   #echo $(pwd)
-  ansible-playbook --connection=local --inventory ./inventory --limit local  playbooks/pb_config.yaml --extra-vars "{'host_hosts': 'local', \
+  ansible-playbook --connection=local --inventory ./inventory/$inventory --limit local  playbooks/pb_config.yaml -vvv --extra-vars "{ \
   'solana_user': '$SOLANA_USER', \
   'validator_name':'$VALIDATOR_NAME', \
-  'secrets_path': '$PATH_TO_VALIDATOR_KEYS', \
-  'flat_path': 'True', \
-  'cluster_environment':'$cluster_environment'\
+  'local_secrets_path': '$PATH_TO_VALIDATOR_KEYS' \
   }"
 
-  ansible-playbook --connection=local --inventory ./inventory --limit local  playbooks/pb_install_monitoring.yaml --extra-vars "@/etc/sv_manager/sv_manager.conf" --extra-vars 'host_hosts=local'
+  ansible-playbook --connection=local --inventory ./inventory/$inventory --limit local  playbooks/pb_install_monitoring.yaml --extra-vars "@/etc/sv_manager/sv_manager.conf"
 
   echo "### Cleanup install folder ###"
   cd ..
   rm -r ./sv_manager
   echo "### Cleanup install folder done ###"
-  echo "### Check your dashboard: https://solana.thevalidators.io/d/e-8yEOXMwerfwe/solana-monitoring-v1-0-preview?&var-server="$VALIDATOR_NAME
+  echo "### Check your dashboard: https://solana.thevalidators.io/d/e-8yEOXMwerfwe/solana-monitoring?&var-server="$VALIDATOR_NAME
 
-  echo Do you want to Uninstall ansible?
+  echo Do you want to UNinstall ansible?
   select yn in "Yes" "No"; do
       case $yn in
           Yes ) $pkg_manager remove ansible --yes; break;;
